# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Career Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  careerScore: number;
  jobsApplied: number;
  skillsTracked: number;
  salaryBenchmark: number;
  marketDemand: string;
  weeklyGrowth: number;
  topRole: string;
  learningProgress: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  timestamp: string;
  icon: string;
}

export interface CareerScore {
  overall: number;
  skills: number;
  experience: number;
  portfolio: number;
  marketFit: number;
  learning: number;
  recommendations: string[];
}

export interface SkillDemand {
  skill: string;
  category: string;
  demandScore: number;
  jobCount: number;
  growthRate: number;
  avgSalaryImpact: number;
  trend: string;
}

export interface TrendingSkill {
  skill: string;
  momentum: number;
  weeklyGrowth: number;
  hotness: string;
}

export interface SkillGapInput {
  currentSkills: string[];
  targetRole: string;
  experienceLevel: string;
}

export interface SkillGapItem {
  skill: string;
  importance: string;
  timeToLearn: string;
}

export interface SkillGapResult {
  targetRole: string;
  matchScore: number;
  missingSkills: SkillGapItem[];
  strongSkills: string[];
  estimatedTimeline: string;
  recommendations: string[];
}

export interface SalaryData {
  role: string;
  location: string;
  experience: string;
  minSalary: number;
  maxSalary: number;
  avgSalary: number;
  currency: string;
  industry: string;
}

export interface SalaryTrend {
  year: number;
  role: string;
  avgSalary: number;
  growth: number;
}

export interface RoleSalary {
  role: string;
  avgSalary: number;
  demand: string;
}

export interface LocationSalary {
  location: string;
  avgSalary: number;
  costOfLivingIndex: number;
}

export interface IndustrySalary {
  industry: string;
  avgSalary: number;
  growth: number;
}

export interface SalaryComparison {
  roles: RoleSalary[];
  locations: LocationSalary[];
  industries: IndustrySalary[];
}

export interface ResumeInput {
  resumeText: string;
  targetRole: string;
  targetIndustry?: string;
}

export interface ResumeImprovement {
  section: string;
  priority: string;
  suggestion: string;
}

export interface ResumeAnalysis {
  id: number;
  atsScore: number;
  overallScore: number;
  targetRole: string;
  analyzedAt: string;
  strengths: string[];
  weaknesses: string[];
  missingKeywords: string[];
  suggestedKeywords: string[];
  improvements: ResumeImprovement[];
  skillsDetected: string[];
  formatScore: number;
  contentScore: number;
  keywordScore: number;
}

export interface SavedResume {
  id: number;
  targetRole: string;
  atsScore: number;
  analyzedAt: string;
}

export interface CoachMessage {
  id: number;
  role: string;
  content: string;
  timestamp: string;
  suggestions?: string[];
}

export interface CoachMessageInput {
  message: string;
  context?: string;
}

export interface RoadmapInput {
  currentRole: string;
  targetRole: string;
  currentSkills?: string[];
  timeframe: string;
  learningStyle?: string;
}

export interface RoadmapPhase {
  phase: number;
  title: string;
  duration: string;
  skills: string[];
  milestone: string;
  completed: boolean;
}

export interface LearningResource {
  title: string;
  type: string;
  platform: string;
  duration: string;
  difficulty: string;
  url?: string;
}

export interface LearningRoadmap {
  id: number;
  title: string;
  targetRole: string;
  timeframe: string;
  completionPercentage: number;
  createdAt: string;
  phases: RoadmapPhase[];
  resources: LearningResource[];
}

export interface JobListing {
  id: number;
  title: string;
  company: string;
  location: string;
  remote: boolean;
  salaryMin: number;
  salaryMax: number;
  skills: string[];
  postedAt: string;
  matchScore: number;
  industry: string;
  experienceLevel: string;
}

export interface HiringCompany {
  name: string;
  industry: string;
  openRoles: number;
  avgSalary: number;
  rating: number;
  remote: boolean;
  topSkills: string[];
}

export interface LocationDemand {
  location: string;
  openings: number;
  growth: number;
}

export interface RoleDemand {
  role: string;
  openings: number;
  percentage: number;
}

export interface IndustryDemand {
  industry: string;
  openings: number;
  growth: number;
}

export interface JobMarketOverview {
  totalOpenings: number;
  remotePercentage: number;
  avgTimeToHire: number;
  demandGrowth: number;
  topLocations: LocationDemand[];
  roleBreakdown: RoleDemand[];
  industryDemand: IndustryDemand[];
}

export interface MarketPrediction {
  skill: string;
  currentDemand: number;
  predictedDemand: number;
  growthRate: number;
  timeframe: string;
  confidence: number;
  insight: string;
}

export interface PortfolioProject {
  id: number;
  title: string;
  description: string;
  skills: string[];
  category: string;
  score: number;
  recruiterReadiness: string;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  complexity: string;
  impact: string;
  createdAt: string;
}

export interface PortfolioProjectInput {
  title: string;
  description: string;
  skills: string[];
  category: string;
  githubUrl?: string;
  liveUrl?: string;
  complexity?: string;
}

export interface PortfolioScore {
  overall: number;
  diversity: number;
  complexity: number;
  presentation: number;
  recruiterReadiness: number;
  totalProjects: number;
  strengths: string[];
  improvements: string[];
  topProject: string;
}

export type GetSkillDemandParams = {
category?: string;
timeframe?: string;
};

export type GetSalaryByRoleParams = {
role?: string;
location?: string;
experience?: string;
};

export type GetJobListingsParams = {
role?: string;
location?: string;
remote?: boolean;
};