# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Data Career Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolio/projects/{id}:
    get:
      operationId: getPortfolioProject
      tags: [portfolio]
      summary: Get a specific portfolio project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioProject"
  /portfolio/projects/add:
    post:
      operationId: addPortfolioProject
      tags: [portfolio]
      summary: Add a new portfolio project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioProjectInput"
      responses:
        "201":
          description: Created portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioProject"
...
components:
  schemas:
...
    PortfolioProject:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        category:
          type: string
        score:
          type: integer
        recruiterReadiness:
          type: string
        githubUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        complexity:
          type: string
        impact:
          type: string
        createdAt:
          type: string
      required: [id, title, description, skills, category, score, recruiterReadiness, complexity, impact, createdAt]
    PortfolioProjectInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        category:
          type: string
        githubUrl:
          type: string
        liveUrl:
          type: string
        complexity:
          type: string
      required: [title, description, skills, category]
...