# lib/api-zod/src/generated/types/portfolioProject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Career Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioProject {
  id: number;
  title: string;
  description: string;
  skills: string[];
  category: string;
  score: number;
  recruiterReadiness: string;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  complexity: string;
  impact: string;
  createdAt: string;
}